Creation of initial Rails application.
echo "source 'https://rubygems.org'" > Gemfile
echo "gem 'rails', '~> 4.2.0'" >> Gemfile
bundle exec rails new . -f

To do this, we created a new Gemfile, with Rails 4.2 as the only
dependency, and ran 'rails new' to create our app.

When the destination path is ".", rails will infer the name of the
application from the current directory name.

The -f flag is used to overwrite the initial Gemfile that we
created.