Adding scaffolding for Comments
rails generate scaffold Comment message:text submission:references parent:references

A comment will have some text, it will relate to one submission, and may
have a parent comment.  As we are scaffolding here, it is not yet clear
to rails that the 'parent' that is being referenced is a comment.  We
will fix that later.