Adding model validations for Submission
We are validating that submissions are associated with a community.  We
are also validating that a submission has a title of at least five
characters and the title is unique.

We are including the validate_url gem in our Gemfile in order to add the
validation that a submission's URL must be present and must be a valid
URL.

We are updating the submissions controller test to deal with the fact
that submission titles must be unique.

We are updating the submissions fixture file to correctly associate the
fixture submissions with the fixture communities.