Adding model validations for Comments
We are validating that a comment must have a message of length at
least 1 character.  We are validating that it must be associated with a
submission.

We are creating the relation between a Comment and its parent (which is
also a Comment).

We are modifying the comments fixtures in order to have them relate to
actual submission fixtures.