Adding the ability for the user to comment on submissions
We add a link on the Submissions show page that goes to the "new" action
of the Comments controller.  We include a parameter indicating which
Submission it is related to.

In the Comments controller we use comment_params to make sure the new
Comment we are creating respects the parameters passed in (right now,
it's just a submission id, but we will use this again for comment
replies).

In the Comments view form partial, we've changed the form field
corresponding to submission_id to be a hidden field (since it will
always be supplied beforehand) and included the title of the submission.