Adding the ability to reply to Comments
We have broken out the display of the message into its own partial.
This partial renders itself recursively, displaying its message and then
rendering its children.

We have added the display of all comments to the bottom of a submission
using this new partial on the submissions show page.  In order to do
this, we have added a method to the submission model to return all "root
comments" (those without parents).

We are adding a "Reply" link to each comment.  This link goes to the
"new" action of the comments controller and sends in the correct parent
and submission id to associate the reply with the original comment.

The form for entering a new comment has been changed to show the parent
comment and has a hidden form field for setting the parent correctly
when submitted.

We have also made it so that when successfully creating a comment the
user is directed back to the submission that was originally commented
on.