Cleaned up the submissions page and made it the default
Made / point to /submissions.  Changed the title to something more
appropriate.  Made the title of each submission a clickable link.

Removed the delete and edit links, and changed the show links to list
the number of comments on the submission.

Since this will be the default page, a link for creating a new community
has been added.