Adding database seed data
Adding the Faker gem in order to easily generate seed data.

Modified db/seeds.rb to create random communities, submissions and
comments.  Because the function for randomness is different between
mysql and sqlite, we need to make that dependent on the name of the
ActiveRecord adaptor.

It should be noted that this random data is not generated
deterministically, so we should remain aware of this when using it for
load testing.

Also note that we continuously check the number of communities,
submissions and comments that already exist in the database.  We do this
because the seed task may be run on multiple app servers that start at
the same time.